chore: rename TopicPlayer as TopicViewer
